refactor: Moved stelae git logic to src/server/api/stelae folder
