+++ src/server/api/mod.rs
@@ -2,4 +2,5 @@
 pub mod routes;
 pub mod serve;
 pub mod state;
+pub mod stelae;
 pub mod versions;
