+++ src/server/api/routes.rs
@@ -3,6 +3,7 @@
     clippy::exit,
     reason = "We exit with 1 error code on any application errors"
 )]
+use std::sync::Arc;
 use std::{process, sync::OnceLock};
 
 use crate::server::api::state;
@@ -11,18 +12,11 @@ use actix_service::ServiceFactory;
 use actix_web::{
     body::MessageBody,
     dev::{ServiceRequest, ServiceResponse},
-    guard, route, web, App, Error, HttpResponse, Responder, Scope,
+    guard, web, App, Error, Scope,
 };
-use serde::Deserialize;
 
-use super::state::App as AppState;
+use super::stelae::get_blob;
 use super::{serve::serve, state::Global, versions::versions};
-use crate::utils::git::{Repo, GIT_REQUEST_NOT_FOUND};
-use crate::utils::http::get_contenttype;
-use crate::utils::paths::clean_path;
-use git2::{self, ErrorCode};
-
-use super::super::errors::HTTPError;
 
 /// Name of the header to guard current documents
 static HEADER_NAME: OnceLock<String> = OnceLock::new();
@@ -80,9 +74,14 @@ pub fn register_app<
         )
         .app_data(web::Data::new(state.clone()));
 
-    app = app
-        .service(web::scope("/_git").service(get_blob))
-        .app_data(web::Data::new(state.clone()));
+    let stelae_data: Arc<dyn Global> = Arc::new(state.clone());
+    app = app.app_data(web::Data::new(stelae_data.clone())).service(
+        web::scope("_stelae").service(
+            web::resource("/{namespace}/{name}")
+                .route(web::get().to(get_blob))
+                .route(web::head().to(get_blob)),
+        ),
+    );
 
     app = register_dynamic_routes(app, state)?;
     Ok(app)
@@ -326,64 +325,4 @@ fn register_dependent_routes(
         cfg.service(actix_scope);
     }
     Ok(())
-}
-
-/// Structure for
-#[derive(Debug, Deserialize)]
-struct Info {
-    /// commit of the repo
-    commitish: String,
-    /// path of the file
-    remainder: Option<String>,
-}
-
-/// Return the content in the stelae archive in the `{namespace}/{name}`
-/// repo at the `commitish` commit at the `remainder` path.
-/// Return 404 if any are not found or there are any errors.
-#[route("/{namespace}/{name}", method = "GET", method = "HEAD")]
-#[tracing::instrument(name = "Retrieving a Git blob", skip(path, data, info))]
-#[expect(
-    clippy::future_not_send,
-    reason = "We don't worry about git2-rs not implementing `Send` trait"
-)]
-async fn get_blob(
-    path: web::Path<(String, String)>,
-    info: web::Query<Info>,
-    data: web::Data<AppState>,
-) -> impl Responder {
-    let (namespace, name) = path.into_inner();
-    let info_struct: Info = info.into_inner();
-    let commitish = info_struct.commitish;
-    let remainder = info_struct.remainder.unwrap_or_default();
-    let archive_path = &data.archive_path;
-    let blob = Repo::find_blob(archive_path, &namespace, &name, &remainder, &commitish);
-    let blob_path = clean_path(&remainder);
-    let contenttype = get_contenttype(&blob_path);
-    match blob {
-        Ok(content) => HttpResponse::Ok().insert_header(contenttype).body(content),
-        Err(error) => blob_error_response(&error, &namespace, &name),
-    }
-}
-
-/// A centralised place to match potentially unsafe internal errors to safe user-facing error responses
-#[expect(clippy::wildcard_enum_match_arm, reason = "Allows _ for enum matching")]
-#[tracing::instrument(name = "Error with Git blob request", skip(error, namespace, name))]
-fn blob_error_response(error: &anyhow::Error, namespace: &str, name: &str) -> HttpResponse {
-    tracing::error!("{error}",);
-    if let Some(git_error) = error.downcast_ref::<git2::Error>() {
-        return match git_error.code() {
-            // TODO: check this is the right error
-            ErrorCode::NotFound => {
-                HttpResponse::NotFound().body(format!("repo {namespace}/{name} does not exist"))
-            }
-            _ => HttpResponse::InternalServerError().body("Unexpected Git error"),
-        };
-    }
-    match error {
-        // TODO: Obviously it's better to use custom `Error` types
-        _ if error.to_string() == GIT_REQUEST_NOT_FOUND => {
-            HttpResponse::NotFound().body(HTTPError::NotFound.to_string())
-        }
-        _ => HttpResponse::InternalServerError().body(HTTPError::InternalServerError.to_string()),
-    }
 }
