+++ src/server/api/stelae/request/mod.rs
@@ -1,0 +1,9 @@
+use serde::Deserialize;
+/// Structure for
+#[derive(Debug, Deserialize)]
+pub struct StelaeQueryData {
+    /// commit of the repo
+    pub commitish: Option<String>,
+    /// path of the file
+    pub remainder: Option<String>,
+}
